fix: remove trailing slash from zarr paths.  (#362)
## Description

This PR adds a small fix to the Anemoi datasets interface that removes a
trailing slash from .zarr/ paths passed to open_dataset. Without this
fix, providing a path like "path/to/dataset.zarr/" results in a
misleading ValueError, even when the dataset exists.

With this change, both
```
open_dataset("path/to/dataset.zarr/")
```
and
```
open_dataset("path/to/dataset.zarr")
```

will correctly resolve and open the dataset and log a warning that the
path has been changed .

## What problem does this change solve?
This is a bugfix. The open_dataset function fails to locate .zarr
datasets when the input path ends with a trailing slash.


***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

Co-authored-by: Florian Pinault <floriankrb@users.noreply.github.com>