feat: Extend xarray coordinate guesser to include pressure as a long_name (#335)
## Description

This extends `DefaultCoordinateGuesser._is_level` to include a match
when the coordinate has a "long_name" of pressure and the appropriate
pressure units.

This also adds some small unit tests for `DefaultCoordinateGuesser`.

## Type of Change

-   [ ] Bug fix (non-breaking change which fixes an issue)
-   [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
-   [ ] Documentation update

## Issue Number

I am using anemoi-datasets with some netcdf files that currently aren't
working as they have a pressure coordinate with a long_name of
"pressure", and this isn't being recognised as a LevelCoordinate. This
is some example ncdump output:
```
	float pressure(pressure) ;
		pressure:axis = "Z" ;
		pressure:units = "Pa" ;
		pressure:long_name = "pressure" ;
		pressure:positive = "down" ;
```
I have extended the coordinate guesser to recognise these as a
LevelCoordinate

## Code Compatibility

-   [x] I have performed a self-review of my code

### Code Performance and Testing

- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I ran the [complete Pytest
test](https://anemoi.readthedocs.io/projects/training/en/latest/dev/testing.html)
suite locally, and they pass

### Dependencies

- [x] I have ensured that the code is still pip-installable after the
changes and runs
- [ ] I have tested that new dependencies themselves are
pip-installable.

<!-- List any new dependencies that are required for this change and the
justification to add them. -->

### Documentation

-   [x] My code follows the style guidelines of this project
- [ ] I have updated the documentation and docstrings to reflect the
changes
- [ ] I have added comments to my code, particularly in
hard-to-understand areas

<!-- Describe any major updates to the documentation -->

## Additional Notes

<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

Co-authored-by: b8raoult <53792887+b8raoult@users.noreply.github.com>