fix(grib-index): match against filename, not full path (#374)
The `--match` argument of the `grib-index` command is supposed to be
used for matching filenames that are inside the specified directories
(passed as `paths`). However, currently, the code tries to match the
full path instead of the filename only. For instance, let's say we want
to match files inside `/path/to/grib/files`. My files have pattern
`lafYYYYmmddHH.grib`. If I try to match with `laf*` it won't work. If I
try with `*.grib` it will.

With this PR, only the filename is checked and therefore also pattern
with the * at the end will work.

Co-authored-by: Florian Pinault <floriankrb@users.noreply.github.com>