fix: deprecated filters in anemoi datasets (#350)
## Description
Deprecate and migrate filters from anemoi-datasets

## What problem does this change solve?
Remove and harmonise filters in anemoi-transform

## What issue or task does this change relate to?
[<!-- link to Issue Number
-->](https://github.com/ecmwf/anemoi-transform/issues/92)

##  Additional notes ##

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***


<!-- readthedocs-preview anemoi-datasets start -->
----
📚 Documentation preview 📚:
https://anemoi-datasets--350.org.readthedocs.build/en/350/

<!-- readthedocs-preview anemoi-datasets end -->

---------

Co-authored-by: ClémentBrochet <95094184+flyIchtus@users.noreply.github.com>
Co-authored-by: Aaron Hopkinson <aaron.hopkinson@ecmwf.int>