fix: save_dataset (#383)
## Description
Recent slack conversations showed that there were issues with the
save_dataset method. It now:

- takes a Dataset (opened anemoi-dataset) as an input instead of a
recipe, because it was confusing to have different configs to open and
to create for most people trying to use it;
- registers metadata so that you can run anemoi-datasets inspect on it;
- fixes the error with the ensemble dim.

## What problem does this change solve?
The fact that people could not inspect created dataset and the ensemble
dim was wrong.

## What issue or task does this change relate to?
NA

##  Additional notes ##


***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

---------

Co-authored-by: Aaron Hopkinson <197336788+aaron-hopkinson@users.noreply.github.com>