feat: refactor dataset create (#379)
## Description

This is a rewrite of the code that parses the recipe and invoke the
sources and filters in the right order. These changes are merely a
simplification of the old version and are 100% backward compatible.
The input family of actions has been refactored such that all the
standalone action files, e.g. concat are consolidated into a single
action.py.

Additional related additions.
- Add facility to have a source as a parameter of an other source if
needed.
- Changes are due to code moving to their own file (i.e repeated_dates).
- A new CLI tool anemoi-datasets recipe was added to check the change.
It can be use to format the YAML and validate it (against a simple JSON
schema), and migrate older YAML files to new syntax. The later is
experimental, and as been tested with very old recipe files. It can be a
good starting point for future migrations, if ever needed.


## What problem does this change solve?

This change is in preparation of adding support for observation in
anemoi-dataset[create].

## What issue or task does this change relate to?
<!-- link to Issue Number -->

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

---------

Co-authored-by: Florian Pinault <floriankrb@users.noreply.github.com>
Co-authored-by: Matthew Chantry <matthew.chantry@ecmwf.int>
Co-authored-by: Marek Jacob <MeraX@users.noreply.github.com>
Co-authored-by: Aaron Hopkinson <197336788+aaron-hopkinson@users.noreply.github.com>