feat: added rolling average on read (#448)
## Description
This allows to compute the rolling average at read time using
`rolling_average` keyword.
We do not want to use this pattern (of loading extra data at read time)
because it is sub-optimal, but this is an easy work-around until we are
able to compute the rolling average when building the dataset.


***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

---------

Co-authored-by: Aaron Hopkinson <197336788+aaron-hopkinson@users.noreply.github.com>
Co-authored-by: Ana Prieto Nemesio <91897203+anaprietonem@users.noreply.github.com>